% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R, R/mongrelfit_methods.R
\name{access_dims}
\alias{access_dims}
\alias{ncategories}
\alias{nsamples}
\alias{ncovariates}
\alias{niter}
\alias{ncategories.mongrelfit}
\alias{nsamples.mongrelfit}
\alias{ncovariates.mongrelfit}
\alias{niter.mongrelfit}
\title{Generic method for accessing model fit dimensions}
\usage{
ncategories(m)

nsamples(m)

ncovariates(m)

niter(m)

\method{ncategories}{mongrelfit}(m)

\method{nsamples}{mongrelfit}(m)

\method{ncovariates}{mongrelfit}(m)

\method{niter}{mongrelfit}(m)
}
\arguments{
\item{m}{An object of class mongrelfit}
}
\value{
integer
}
\description{
Generic method for accessing model fit dimensions
}
\details{
An alternative approach to accessing these dimensions is to 
  access them directly from the mongrelfit object using list indexing. 
* \code{ncategories} is equivalent to \code{m$D}
* \code{nsamples} is equivalent to \code{m$N}
* \code{ncovariates} is equivalent to \code{m$Q}
}
\examples{
\dontrun{
m <- mongrel(Y, X)
ncategories(m)
nsamples(m)
ncovariates(m)
}
}
