% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_utils.R
\name{add_array_dim}
\alias{add_array_dim}
\title{Add 1 dimension to array}
\usage{
add_array_dim(a, d)
}
\arguments{
\item{a}{array object}

\item{d}{dimension to add (<= length(dim(a))+1)}
}
\value{
array
}
\description{
Add 1 dimension to array
}
\examples{
x <- matrix(1:6, 3, 2)
add_array_dim(a, 1)
}
