% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray_utils.R
\name{check_dims}
\alias{check_dims}
\title{Check vector/matrix/data.frame for expected dimensions or throw error}
\usage{
check_dims(x, d, par)
}
\arguments{
\item{x}{object to check}

\item{d}{expected dimensions}

\item{par}{character name of x (for error message)}
}
\value{
nothing if no error, otherwise throws error
}
\description{
Check vector/matrix/data.frame for expected dimensions or throw error
}
\examples{
y <- c(1,3,4)
check_dims(y, 3, "y")
}
