% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check duplicates}
\usage{
check_duplicates(data)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}
}
\value{
A list in which each component is a vector of
observation indexes. The first index in each list is the
index of the exemplar defining that list, and any remaining indexes are the associated members, within radius of the exemplar.
}
\description{
Check duplicates by giving a one-to-one mapping from unique observations to rows of a data matrix
}
