% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_methods.R
\name{coef.mongrelfit}
\alias{coef.mongrelfit}
\title{Return regression coefficients of mongrelfit object}
\usage{
\method{coef}{mongrelfit}(m, use_names = TRUE)
}
\arguments{
\item{m}{an object of class mongrelfit}

\item{use_names}{if column and row names were passed for Y and X in 
call to \code{\link{fit_mongrel}}, should these names be applied to output 
array.}
}
\value{
Array of dimension (D-1) x Q x iter
}
\description{
Returned as array of dimension (D-1) x Q x iter.
}
\examples{
\dontrun{
fit <- fit_mongrel(Y, X)
coef(fit)
}
}
