% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{coef.pibblefit}
\alias{coef.pibblefit}
\title{Return regression coefficients of pibblefit object}
\usage{
\method{coef}{pibblefit}(object, ...)
}
\arguments{
\item{object}{an object of class pibblefit}

\item{...}{other options passed to coef.pibblefit (see details)}
}
\value{
Array of dimension (D-1) x Q x iter
}
\description{
Returned as array of dimension (D-1) x Q x iter (if in ALR or ILR) otherwise
DxQxiter (if in proportions or clr).
}
\details{
Other arguments:
\itemize{
\item `use_names` if column and row names were passed for Y and X in 
call to \code{\link{pibble}}, should these names be applied to output 
array. 
}
}
\examples{
\dontrun{
fit <- pibble(Y, X)
coef(fit)
}
}
