% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_mp_status.R
\name{compiled_with_openmp}
\alias{compiled_with_openmp}
\title{Check if stray was compiled with OpenMP}
\usage{
compiled_with_openmp()
}
\value{
boolean
}
\description{
hopefully helpful when first installing stray
}
