% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthus_transform_wrapper.R
\name{convert_orthus_covariance}
\alias{convert_orthus_covariance}
\alias{oilrvar2ilrvar}
\alias{oilrvar2clrvar}
\alias{oclrvar2ilrvar}
\alias{oalrvar2clrvar}
\alias{oclrvar2alrvar}
\alias{oalrvar2alrvar}
\alias{oalrvar2ilrvar}
\alias{oilrvar2alrvar}
\title{Convert orthus covariance matricies between representations}
\usage{
oilrvar2ilrvar(Sigma, s, V1, V2)

oilrvar2clrvar(Sigma, s, V)

oclrvar2ilrvar(Sigma, s, V)

oalrvar2clrvar(Sigma, s, d1)

oclrvar2alrvar(Sigma, s, d2)

oalrvar2alrvar(Sigma, s, d1, d2)

oalrvar2ilrvar(Sigma, s, d1, V2)

oilrvar2alrvar(Sigma, s, V1, d2)
}
\arguments{
\item{Sigma}{covariance matrix arrat in specified transformed space 
(dim(Sigma)[3]=iter)}

\item{s}{first s rows and colums of Sigma are transformed}

\item{V1}{ILR contrast matrix of basis Sigma is already in}

\item{V2}{ILR contrast matrix of basis Sigma is desired in}

\item{V}{ILR contrast matrix (i.e., transformation matrix of ILR)}

\item{d1}{alr reference element Sigma is already expressed with respec to}

\item{d2}{alr reference element Sigma is to be expressed with respect to}
}
\value{
matrix
}
\description{
Convert orthus covariance matricies between representations
}
