% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_b}
\alias{data_b}
\title{A bimodal dataset with a micro cluster}
\format{A data frame with 2003 rows and 3 variables:
 \describe{
\item{x }{numerical variable}
\item{y }{numerical variable}
\item{type}{Type of a data point : Typical or Outlier}}}
\usage{
data_b
}
\description{
A bivariate dataset with two typical classes and a micor cluster
}
\keyword{datasets}
