% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_c}
\alias{data_c}
\title{A dataset with local anomalies and  micro clusters}
\format{A data frame with 1009 rows and 3 variables:
 \describe{
\item{x }{numerical variable}
\item{y }{numerical variable}
\item{type}{Type of a data point : Typical or Outlier}}}
\usage{
data_c
}
\description{
A bivariate dataset with local anomalies and two  micro clusters
}
\keyword{datasets}
