% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_d}
\alias{data_d}
\title{A wheel dataset with  two inliers}
\format{A data frame with 1002 rows and 3 variables:
 \describe{
\item{x }{numerical variable}
\item{y }{numerical variable}
\item{type}{Type of a data point : Typical or Outlier}}}
\usage{
data_d
}
\description{
A bivariate dataset with  two inliers. The inliers are very close to one another
}
\keyword{datasets}
