% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_e}
\alias{data_e}
\title{A bimodal dataset with  an inlier}
\format{A data frame with 2001 rows and 3 variables:
 \describe{
\item{x }{numerical variable}
\item{y }{numerical variable}
\item{type}{Type of a data point : Typical or Outlier}}}
\usage{
data_e
}
\description{
A bimodal dataset with  an inlier. one typical class is a very dense  cluster,
}
\keyword{datasets}
