% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_f}
\alias{data_f}
\title{A dataset with  an outlier}
\format{A data frame with 2001 rows and 3 variables:
 \describe{
\item{x }{numerical variable}
\item{y }{numerical variable}
\item{type}{Type of a data point : Typical or Outlier}}}
\usage{
data_f
}
\description{
A dataset with  an outlier. The typical class is a very dense cluster.
}
\keyword{datasets}
