% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Detect Anomalies Based on the Maximum k-Nearest Neighbour Differences.}
\usage{
get_outliers(data, memberLists, alpha = 0.05, l = 0.2)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}

\item{memberLists}{A list following the structure of the output to \code{\link[HDoutliers]{getHDmembers}}, in
which each component is a vector of observation indexes. The first index in each list is the index of the
exemplar representing that list, and any remaining indexes are the associated members, considered
'close to' the exemplar.}

\item{alpha}{Threshold for determining the cutoff for outliers. Observations are considered outliers
outliers if they fall in the (1- alpha) tail of the distribution of the nearest-neighbor distances
between exemplars.}

\item{l}{A parameter the user can choose to obtain a trade off between the false positives and
the false negatives.}
}
\value{
The indexes of the observations determined to be outliers.
}
\description{
Detect anomalies based on the maximum k-Nearest Neighbour differences. This is a modification of
\code{\link[HDoutliers]{getHDoutliers}}.
}
\seealso{
\code{\link[HDoutliers]{getHDmembers}},  \code{\link{find_HDoutliers}}
}
