% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{hdr_outliers}
\alias{hdr_outliers}
\title{Find anomalies using high density regions of the irst two principal components of the high dimensional data set}
\usage{
hdr_outliers(data)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}
}
\value{
The indexes of the observations determined to be outliers.
}
\description{
Find anomalies using high density regions of the irst two principal components of the high dimensional data set
}
