% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hessVectorProd}
\alias{hessVectorProd}
\title{Hessian Vector Product using Finite Differences}
\usage{
hessVectorProd(Y, upsilon, ThetaX, K, A, eta, v, r)
}
\description{
Hessian Vector Product using Finite Differences
}
