% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmvgamma}
\alias{lmvgamma}
\title{Log of Multivarate Gamma Function - Gamma_p(a)}
\usage{
lmvgamma(a, p)
}
\description{
Log of Multivarate Gamma Function - Gamma_p(a)
}
\references{
https://en.wikipedia.org/wiki/Multivariate_gamma_function
}
