% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmvgamma_deriv}
\alias{lmvgamma_deriv}
\title{Derivative of Log of Multivariate Gamma Function - Gamma_p(a)}
\usage{
lmvgamma_deriv(a, p)
}
\description{
Derivative of Log of Multivariate Gamma Function - Gamma_p(a)
}
