% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikMaltipooCollapsed}
\alias{loglikMaltipooCollapsed}
\title{Calculate log posterior (up to proportionality constant) of Maltipoo model}
\usage{
loglikMaltipooCollapsed(Y, upsilon, Theta, X, K, U, eta, ell,
  sylv = FALSE)
}
\arguments{
\item{Y}{D x N matrix of counts}

\item{upsilon}{(must be > D)}

\item{Theta}{D-1 x Q matrix the prior mean for regression coefficients}

\item{X}{Q x N matrix of covariates}

\item{U}{a PQxQ matrix of stacked variance components}
}
\description{
Calculate log posterior (up to proportionality constant) of Maltipoo model
}
