% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallard-data.R
\docType{data}
\name{mallard}
\alias{mallard}
\title{Data from Silverman et al. (2018) Microbiome}
\format{An object of class \code{\link[phyloseq]{phyloseq-class}}}
\usage{
data(mallard)
}
\description{
High Resolution (hourly and daily) sampling of 4 in vitro artificial gut models
with many technical replicates to identify technical variation.
}
\references{
Silverman et al. "Dynamic linear models guide design and 
  analysis of microbiota studies within artificial human guts". 
  Microbiome 2018 6:202
}
\keyword{datasets}
