% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallard_family-data.R
\docType{data}
\name{mallard_family}
\alias{mallard_family}
\title{Data from Silverman et al. (2018) Microbiome}
\format{
An object of class \code{\link[phyloseq]{phyloseq-class}}
}
\usage{
data(mallard_family)
}
\description{
High Resolution (hourly and daily) sampling of 4 in vitro artificial gut models
with many technical replicates to identify technical variation.
}
\details{
This data is at the family level and processed as in Silverman et al. 2018. Data at the sequence
variant level without preprocessing is given in \code{\link{mallard}}
}
\references{
Silverman et al. "Dynamic linear models guide design and 
  analysis of microbiota studies within artificial human guts". 
  Microbiome 2018 6:202
}
