% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maltipoofit_s3.R
\name{maltipoofit}
\alias{maltipoofit}
\title{Create maltipoofit object}
\usage{
maltipoofit(
  D,
  N,
  Q,
  P,
  coord_system,
  iter = NULL,
  alr_base = NULL,
  ilr_base = NULL,
  Eta = NULL,
  Lambda = NULL,
  Sigma = NULL,
  Sigma_default = NULL,
  Y = NULL,
  X = NULL,
  upsilon = NULL,
  Theta = NULL,
  Xi = NULL,
  Xi_default = NULL,
  Gamma = NULL,
  init = NULL,
  ellinit = NULL,
  names_categories = NULL,
  names_samples = NULL,
  names_covariates = NULL,
  VCScale = NULL,
  U = NULL
)
}
\arguments{
\item{D}{number of multinomial categories}

\item{N}{number of samples}

\item{Q}{number of covariates}

\item{P}{number of variance components}

\item{coord_system}{coordinate system objects are represented in (options 
include "alr", "clr", "ilr", and "proportions")}

\item{iter}{number of posterior samples}

\item{alr_base}{integer category used as reference 
(required if coord_system=="alr")}

\item{ilr_base}{(D x D-1) contrast matrix (required if coord_system=="ilr")}

\item{Eta}{Array of samples of Eta}

\item{Lambda}{Array of samples of Lambda}

\item{Sigma}{Array of samples of Sigma (null if coord_system=="proportions")}

\item{Sigma_default}{Array of samples of Sigma in alr base D, used if 
coord_system=="proportions"}

\item{Y}{DxN matrix of observed counts}

\item{X}{QxN design matrix}

\item{upsilon}{scalar prior dof of inverse wishart prior}

\item{Theta}{prior mean of Lambda}

\item{Xi}{Matrix of prior covariance for inverse wishart 
(null if coord_system=="proportions")}

\item{Xi_default}{Matrix of prior covariance for inverse wishart in alr 
base D (used if coord_system=="proportions")}

\item{Gamma}{QxQ covariance matrix prior for Lambda}

\item{init}{matrix initial guess for Lambda used for optimization}

\item{ellinit}{P vector initialization values for ell for optimization}

\item{names_categories}{character vector}

\item{names_samples}{character vector}

\item{names_covariates}{character vector}

\item{VCScale}{scale factors (delta) for variance components}

\item{U}{a PQ x Q matrix of stacked variance components (each of dimension Q x Q)}
}
\value{
object of class maltipoofit
}
\description{
Create maltipoofit object
}
\seealso{
\code{\link{maltipoo}}
}
