% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mongrel.R
\name{mongrel_fit}
\alias{mongrel_fit}
\alias{mongrel}
\alias{refit.mongrelfit}
\title{Interface to fit mongrel models}
\usage{
mongrel(Y = NULL, X = NULL, upsilon = NULL, Theta = NULL,
  Gamma = NULL, Xi = NULL, init = NULL, pars = c("Eta", "Lambda",
  "Sigma"), ...)

\method{refit}{mongrelfit}(m, pars = c("Eta", "Lambda", "Sigma"), ...)
}
\arguments{
\item{Y}{D x N matrix of counts (if NULL uses priors only)}

\item{X}{Q x N matrix of covariates (design matrix) (if NULL uses priors only, must
be present to sample Eta)}

\item{upsilon}{dof for inverse wishart prior (numeric must be > D)
(default: D+3)}

\item{Theta}{(D-1) x Q matrix of prior mean for regression parameters
(default: matrix(0, D-1, Q))}

\item{Gamma}{QxQ prior covariance matrix
(default: diag(Q))}

\item{Xi}{(D-1)x(D-1) prior covariance matrix
(default: ALR transform of diag(1)*(upsilon-D-2)/2 - this is
essentially iid on "base scale" using Aitchison terminology)}

\item{init}{(D-1) x Q initialization for regression parameters for
optimization}

\item{pars}{character vector of posterior parameters to return}

\item{...}{argumenst passed to \code{\link{optimMongrelCollapsed}} and
\code{\link{uncollapseMongrelCollapsed}}}

\item{m}{object of class mongrelfit}
}
\value{
an object of class mongrelfit
}
\description{
This function is largely a more user friendly wrapper around
\code{\link{optimMongrelCollapsed}} and
\code{\link{uncollapseMongrelCollapsed}}.
See details for model specification.
Notation: \code{N} is number of samples,
\code{D} is number of multinomial categories, \code{Q} is number
of covariates, \code{iter} is the number of samples of \code{eta} (e.g.,
the parameter \code{n_sample} in the function
\code{\link{optimMongrelCollapsed}})
}
\details{
the full model is given by:
\deqn{Y_j ~ Multinomial(Pi_j)}
\deqn{Pi_j = Phi^{-1}(Eta_j)}
\deqn{Eta ~ MN_{D-1 x N}(Lambda*X, Sigma, I_N)}
\deqn{Lambda ~ MN_{D-1 x Q}(Theta, Sigma, Gamma)}
\deqn{Sigma ~ InvWish(upsilon, Xi)}
Where A = (I_N + X * Gamma * X')^{-1}, K^{-1} = Xi is a (D-1)x(D-1)
covariance matrix, Gamma is a Q x Q covariance matrix, and Phi^{-1} is
ALRInv_D transform.

Default behavior is to use MAP estimate for uncollaping collapsed mongrel
model if laplace approximation is not preformed.
}
\examples{
sim <- mongrel_sim()
fit <- mongrel(sim$Y, sim$X)
}
\seealso{
\code{\link{mongrel_transforms}} provide convenience methods for
transforming the representation of mongrelfit objects (e.g., conversion to
proportions, alr, clr, or ilr coordinates.)

\code{\link{access_dims}} provides convenience methods for accessing
dimensions of mongrelfit object

Generic functions including \code{\link[=summary.mongrelfit]{summary}},
\code{\link[=print.mongrelfit]{print}},
\code{\link[=coef.mongrelfit]{coef}},
\code{\link[=as.list.mongrelfit]{as.list}}, and
\code{\link[=predict.mongrelfit]{predict}}

Plotting functions provided by \code{\link[=plot.mongrelfit]{plot}}
and \code{\link{ppc}} (posterior predictive checks)
}
