% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel.R
\docType{package}
\name{mongrel_package}
\alias{mongrel_package}
\alias{mongrel_package-package}
\title{mongrel: Fitting and Analysis of Multinomial Logistic Normal Regression Models}
\description{
Provides methods for fitting and inspection of Bayesian Multinomial 
 Logistic Normal Regression Models ("Mongrel Models") using MAP estimation 
 (with the ADAM optimizer) and Laplace Approximation. Key functionality is 
 implemented in C++ for scalability.
}
