% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_sim.R
\name{mongrel_sim}
\alias{mongrel_sim}
\title{Simulate simple mongrel dataset and priors (for testing)}
\usage{
mongrel_sim(D = 10, N = 30, Q = 2, use_names = TRUE,
  true_priors = FALSE)
}
\arguments{
\item{D}{number of multinomial categories}

\item{N}{number of samples}

\item{Q}{number of covariates (first one is an intercept, must be > 1)}

\item{use_names}{should samples, covariates, and categories be named}

\item{true_priors}{should Xi and upsilon be chosen to have mean at true 
simulated value}
}
\value{
list
}
\description{
Simulate simple mongrel dataset and priors (for testing)
}
\examples{
sim <- mongrel_sim()
}
