% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_methods.R
\name{mongrel_tidy_samples}
\alias{mongrel_tidy_samples}
\title{Convert mongrel samples of Eta Lambda and Sigma to tidy format}
\usage{
mongrel_tidy_samples(m, use_names = FALSE)
}
\arguments{
\item{m}{an object of class mongrelfit}

\item{use_names}{should dimension indicies be replaced by
dimension names if provided in data used to fit mongrel model.}
}
\value{
tibble
}
\description{
Combines them all into a single tibble, see example for formating and 
column headers. Primarily designed to be used by 
\code{\link{summary.mongrelfit}}.
}
\examples{
sim <- mongrel_sim()
attach(sim)
fit <- fit_mongrel(Y, X)
fit_tidy <- mongrel_tidy_samples(fit, use_names=TRUE)
head(fit_tidy)
}
