% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray_utils.R
\name{name.mongrelfit}
\alias{name.mongrelfit}
\title{S3 for mongrelfit apply names to mongrel object}
\usage{
\method{name}{mongrelfit}(m, ...)
}
\arguments{
\item{m}{object of class mongrelfit}

\item{...}{currently ignored}
}
\value{
object of class mongrelfit
}
\description{
S3 for mongrelfit apply names to mongrel object
}
