% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray_utils.R
\name{name.orthusfit}
\alias{name.orthusfit}
\title{S3 for orthusfit apply names to orthusfit object}
\usage{
\method{name}{orthusfit}(m, ...)
}
\arguments{
\item{m}{object of class orthusfit}

\item{...}{currently ignored}
}
\value{
object of class orthusfit
}
\description{
To avoid confusion, assigned default names to multinomial categories (c1 etc...) 
and zdimensions (z1 etc...)
}
