% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray_utils.R
\name{name.pibblefit}
\alias{name.pibblefit}
\title{S3 for pibblefit apply names to pibblefit object}
\usage{
\method{name}{pibblefit}(m, ...)
}
\arguments{
\item{m}{object of class pibblefit}

\item{...}{currently ignored}
}
\value{
object of class pibblefit
}
\description{
S3 for pibblefit apply names to pibblefit object
}
