% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R
\name{name}
\alias{name}
\title{Generic method for applying names to an object}
\usage{
name(m, ...)
}
\arguments{
\item{m}{object}

\item{...}{other arguments to be passed}
}
\value{
object of same class but with names applied to dimensions
}
\description{
Intended to be called internally by package
}
