% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimMMTC}
\alias{optimMMTC}
\title{define optimization function}
\usage{
optimMMTC(Y, upsilon, ThetaX, K, A, etainit, iter = 2000L,
  eigvalthresh = -0.001, calcGradHess = TRUE)
}
\description{
define optimization function
}
