% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthus_transform_wrapper.R
\name{orthus_lr_transforms}
\alias{orthus_lr_transforms}
\alias{oglr}
\alias{oglrInv}
\alias{oalr}
\alias{oalrInv}
\alias{oilr}
\alias{oilrInv}
\alias{oclr}
\alias{oclrInv}
\title{Log-Ratio transforms for orthus objects}
\usage{
oglr(x, s, V)

oglrInv(x, s, V)

oalr(x, s, d = NULL)

oalrInv(y, s, d = NULL)

oilr(x, s, V = NULL)

oilrInv(y, s, V = NULL)

oclr(x, s)

oclrInv(x, s)
}
\arguments{
\item{x}{orthus data array (e.g., first s rows are multinomial parameters or log-ratios)}

\item{s}{first s rows of x are transformed}

\item{V}{transformation matrix (defines transform)}

\item{d}{for ALR, which component (integer position) to take as reference
(default is ncol(x)) for alrInv corresponds to column position in untransformed
matrix.}

\item{y}{orthus data array (e.g., first s rows are multinomial parameters or log-ratios)}
}
\description{
Log-Ratio transforms for orthus objects
}
