% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthus_sim.R
\name{orthus_sim}
\alias{orthus_sim}
\title{Simulate simple orthus dataset and priors (for testing)}
\usage{
orthus_sim(
  D = 10,
  P = 10,
  N = 30,
  Q = 2,
  use_names = TRUE,
  true_priors = FALSE
)
}
\arguments{
\item{D}{number of multinomial categories}

\item{P}{number of dimensions of second dataset Z}

\item{N}{number of samples}

\item{Q}{number of covariates (first one is an intercept, must be > 1)}

\item{use_names}{should samples, covariates, and categories be named}

\item{true_priors}{should Xi and upsilon be chosen to have mean at true 
simulated value}
}
\value{
list
}
\description{
Simulate simple orthus dataset and priors (for testing)
}
\examples{
sim <- orthus_sim()
}
