% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ped_data}
\alias{ped_data}
\title{Dataset with pedestrian counts}
\format{A data frame with 33024 rows and 5 variables:
 \describe{
\item{Sensor }{Sensor location}
\item{Date_Time }{Time and date }
\item{Date}{Date}
\item{Time}{Time}
\item{Count}{Pedestrian count}}}
\usage{
ped_data
}
\description{
A dataset with  hourly pedestrian counts at 43 locations in the
city Melbourne, australia, from 1 December, 2018 to 1, January, 2019.
}
\keyword{datasets}
