% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_plotting.R
\name{ppc.mongrelfit}
\alias{ppc.mongrelfit}
\title{Visualization of Posterior Predictive Check of fit model}
\usage{
\method{ppc}{mongrelfit}(m, iter = 50)
}
\arguments{
\item{m}{an object of class mongrelfit}

\item{iter}{number of samples from posterior predictive distribution to plot
(currently must be <= m$iter)}
}
\value{
ggplot object
}
\description{
Visualization of Posterior Predictive Check of fit model
}
\examples{
\dontrun{
fit <- mongrel(Y, X)
ppc(fit)
}
}
