% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R
\name{ppc}
\alias{ppc}
\title{Generic method for visualizing posterior predictive checks}
\usage{
ppc(m, ...)
}
\arguments{
\item{m}{object}

\item{...}{other arguments passed that control visualization}
}
\description{
Generic method for visualizing posterior predictive checks
}
