% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{predict.mongrelfit}
\alias{predict.mongrelfit}
\title{Predict response from new data}
\usage{
\method{predict}{mongrelfit}(object, newdata = NULL,
  response = "LambdaX", size = NULL, use_names = TRUE,
  summary = FALSE, iter = NULL, ...)
}
\arguments{
\item{object}{An object of class mongrelfit}

\item{newdata}{An optional matrix for which to evaluate predictions. If NULL
(default), the original data of the model is used.}

\item{response}{Options = "LambdaX":Mean of regression, "Eta", "Y": counts}

\item{size}{the number of counts per sample if response="Y" (as vector or matrix), 
default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise
uses median colsums of m$Y as default. If passed as a matrix should have dimensions
ncol(newdata) x iter.}

\item{use_names}{if TRUE apply names to output}

\item{summary}{if TRUE, posterior summary of predictions are returned rather
than samples}

\item{iter}{number of iterations to return if NULL uses object$iter}

\item{...}{other arguments passed to summarise_posterior}
}
\value{
(if summary==FALSE) array D x N x iter; (if summary==TRUE) 
tibble with calculated posterior summaries
}
\description{
Predict response from new data
}
\details{
currently only implemented for mongrelfit objects in coord_system "default"
"alr", or "ilr".
}
\examples{
sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)
}
