% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{print.mongrelfit}
\alias{print.mongrelfit}
\title{Print dimensions and coordinate system information for mongrelfit object.}
\usage{
\method{print}{mongrelfit}(x, summary = FALSE, ...)
}
\arguments{
\item{x}{an object of class mongrelfit}

\item{summary}{if true also calculates and prints summary}

\item{...}{other arguments to pass to summary function}
}
\description{
Print dimensions and coordinate system information for mongrelfit object.
}
\examples{
\dontrun{
fit <- pibble(Y, X)
print(fit)
}
}
\seealso{
\code{\link{summary.mongrelfit}} summarizes posterior intervals
}
