% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{print.orthusfit}
\alias{print.orthusfit}
\title{Print dimensions and coordinate system information for orthusfit object.}
\usage{
\method{print}{orthusfit}(x, summary = FALSE, ...)
}
\arguments{
\item{x}{an object of class orthusfit}

\item{summary}{if true also calculates and prints summary}

\item{...}{other arguments to pass to summary function}
}
\description{
Print dimensions and coordinate system information for orthusfit object.
}
\examples{
\dontrun{
fit <- orthus(Y, Z, X)
print(fit)
}
}
\seealso{
\code{\link{summary.orthusfit}} summarizes posterior intervals
}
