% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_transforms.R
\name{quad_form_array}
\alias{quad_form_array}
\title{Simple Quad form along array's 3rd dimension}
\usage{
quad_form_array(a, V)
}
\arguments{
\item{a}{array of dimension d1 x d1 x d2}

\item{V}{matrix of dimension v1 x d1}
}
\value{
array of dimension v1 x v1 x d2
}
\description{
Simple Quad form along array's 3rd dimension
}
\details{
for i in 1:d2 computes \code{V times a[,,i] times t(V)}
}
