% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R
\name{refit}
\alias{refit}
\title{Generic method for fitting model from passed model fit object}
\usage{
refit(m, ...)
}
\arguments{
\item{m}{object}

\item{...}{other arguments passed that control fitting}
}
\value{
object of the same class as \code{m}
}
\description{
Generic method for fitting model from passed model fit object
}
