% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maltipoofit_s3.R
\name{req.maltipoo}
\alias{req.maltipoo}
\title{require elements to be non-null in mongrelfit or throw error}
\usage{
\method{req}{maltipoo}(m, r)
}
\arguments{
\item{m}{object}

\item{r}{vector of elements to test for}
}
\description{
require elements to be non-null in mongrelfit or throw error
}
