% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrelfit_s3.R
\name{req.mongrelfit}
\alias{req.mongrelfit}
\title{require elements to be non-null in mongrelfit or throw error}
\usage{
req.mongrelfit(m, r)
}
\arguments{
\item{m}{object}

\item{r}{vector of elements to test for}
}
\description{
require elements to be non-null in mongrelfit or throw error
}
