% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_s3.R
\name{req.orthusfit}
\alias{req.orthusfit}
\title{require elements to be non-null in orthusfit or throw error}
\usage{
\method{req}{orthusfit}(m, r)
}
\arguments{
\item{m}{object}

\item{r}{vector of elements to test for}
}
\description{
require elements to be non-null in orthusfit or throw error
}
