% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrelfit_methods.R
\name{sample_prior.mongrelfit}
\alias{sample_prior.mongrelfit}
\title{S3 object for mongrelfit object to sample from prior}
\usage{
\method{sample_prior}{mongrelfit}(m, n_sample = 2000, pars = c("Eta",
  "Lambda", "Sigma"), use_names = TRUE)
}
\arguments{
\item{m}{object of class mongrelfit}

\item{n_sample}{number of samples to produce}

\item{pars}{parameters to sample}

\item{use_names}{should names be used if available}
}
\description{
Note this can be used to sample from prior and then predict can
be called to get counts or LambdaX (\code{\link{predict.mongrelfit}})
}
\details{
Could be greatly speed up in the future if needed by sampling
directly from cholesky form of inverse wishart (currently implemented as 
header in this library - see MatDist.h).
}
