% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R
\name{sample_prior}
\alias{sample_prior}
\title{Generic method for sampling from prior distribution of object}
\usage{
sample_prior(m, n_sample = 2000, ...)
}
\arguments{
\item{m}{object}

\item{...}{other arguments to be passed}

\item{n_samples}{number of samples to produce}
}
\value{
object of the same class
}
\description{
Generic method for sampling from prior distribution of object
}
