% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray_transforms.R
\name{store_coord}
\alias{store_coord}
\alias{reapply_coord}
\title{Holds information on coordinates system to later be reapplied}
\usage{
store_coord(m)

reapply_coord(m, l)
}
\arguments{
\item{m}{object of class mongrelfit}

\item{l}{object returned by function \code{store_coord}}
}
\value{
\code{store_coord} list with important information to identify c
 coordinate system of mongrelfit object. \code{reapply_coord} mongrelfit object
 in coordinate system previously stored.
}
\description{
\code{store_coord} stores coordinate information for mongrelfit object
and can be reapplied with function \code{reapply_coord}. Some coordinate
systems are not useful for computation and this makes it simple keep 
returned object from computations in the same coordinate system as the input. 
(Likely most useful inside of a package)
}
