% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\docType{package}
\name{stray_package}
\alias{stray_package}
\alias{stray_package-package}
\title{stray: Fitting and Analysis of Multinomial Logistic Normal  Models}
\description{
Provides methods for fitting and inspection of Bayesian Multinomial 
 Logistic Normal Models using MAP estimation 
 (with the ADAM optimizer) and Laplace Approximation. Key functionality is 
 implemented in C++ for scalability.
}
