% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray_transforms.R
\name{stray_transforms}
\alias{stray_transforms}
\alias{to_proportions}
\alias{to_alr}
\alias{to_ilr}
\alias{to_clr}
\title{Transform Fit Stray Parameters to other representations}
\usage{
to_proportions(m)

to_alr(m, d)

to_ilr(m, V = NULL)

to_clr(m)
}
\arguments{
\item{m}{object of class mongrelfit (e.g., output of \code{\link{pibble}})}

\item{d}{(integer) multinomial category to take as new alr reference}

\item{V}{(matrix) contrast matrix for ILR basis to transform into to (defaults to 
\code{driver::create_default_ilr_base(D)})}
}
\value{
mongrelfit object
}
\description{
These are a collection of convenience functions for transforming
stray fit objects to a number of different representations including
ILR bases, CLR coordinates, ALR coordinates, and proportions.
}
\details{
Note: that there is a degeneracy of representations for a covariance 
matrix represented in terms of proportions. As such the function 
\code{to_proportions} does not attempt to transform parameters Sigma
or prior Xi and instead just removes them from the mongrelfit object returned.
}
\examples{
\dontrun{
m <- pibble(Y, X)
m.prop <- to_proportions(m)
# convert back to default coordinates (alr with D-th part as reference)
m <- to_alr(m.prop, ncategories(m))
V <- driver::create_default_ilr_base(ncategories(m))
m.ilr <- to_ilr(m, V)
m.clr <- to_clr(m)
}
}
