% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{summary.pibblefit}
\alias{summary.pibblefit}
\title{Summarise pibblefit object and print posterior quantiles}
\usage{
\method{summary}{pibblefit}(
  object,
  pars = NULL,
  use_names = TRUE,
  as_factor = FALSE,
  gather_prob = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class pibblefit}

\item{pars}{character vector (default: c("Eta", "Lambda", "Sigma"))}

\item{use_names}{should summary replace dimension indices with pibblefit 
names if names Y and X were named in call to \code{\link{pibble}}}

\item{as_factor}{if use_names and as_factor then returns names as factors 
(useful for maintaining orderings when plotting)}

\item{gather_prob}{if TRUE then prints quantiles in long format rather than 
wide (useful for some plotting functions)}

\item{...}{other expressions to pass to summarise (using name 'val' unquoted is 
probably what you want)}
}
\description{
Default calculates median, mean, 50\% and 95\% credible interval
}
\examples{
\dontrun{
fit <- pibble(Y, X)
summary(fit, pars="Eta", median = median(val))

# Some later functions make use of precomputation
fit$summary <- summary(fit)
}
}
