% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{use_KNN_maxdiff}
\alias{use_KNN_maxdiff}
\title{Advanced HDoutliers}
\usage{
use_KNN_maxdiff(data, alpha = 0.01, k = 10)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}

\item{alpha}{Threshold for determining the cutoff for outliers. Observations are considered
outliers outliers if they fall in the \eqn{(1- alpha)} tail of the distribution of the nearest-neighbor
distances between exemplars.}

\item{k}{Number of neighbours considered.}
}
\value{
The indexes of the observations determined to be outliers.
}
\description{
Advanced HDoutliers
}
