% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassetfit_s3.R
\name{verify.bassetfit}
\alias{verify.bassetfit}
\title{Simple verification of passed bassetfit object}
\usage{
\method{verify}{bassetfit}(m, ...)
}
\arguments{
\item{m}{an object of class bassetfit}

\item{...}{not used}
}
\value{
throws error if any verification tests fail
}
\description{
Simple verification of passed bassetfit object
}
