% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maltipoofit_s3.R
\name{verify.maltipoofit}
\alias{verify.maltipoofit}
\title{Simple verification of passed multipoo object}
\usage{
\method{verify}{maltipoofit}(m, ...)
}
\arguments{
\item{m}{an object of class multipoo}

\item{...}{not used}
}
\value{
throws error if any verification tests fail
}
\description{
Simple verification of passed multipoo object
}
