% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrelfit_s3.R
\name{verify.mongrelfit}
\alias{verify.mongrelfit}
\title{Simple verification of passed mongrelfit object}
\usage{
\method{verify}{mongrelfit}(m, ...)
}
\arguments{
\item{m}{an object of class mongrelfit}

\item{...}{not used}
}
\value{
throws error if any verification tests fail
}
\description{
Simple verification of passed mongrelfit object
}
