% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_s3.R
\name{verify.orthusfit}
\alias{verify.orthusfit}
\title{Simple verification of passed orthusfit object}
\usage{
\method{verify}{orthusfit}(m, ...)
}
\arguments{
\item{m}{an object of class orthusfit}

\item{...}{not used}
}
\value{
throws error if any verification tests fail
}
\description{
Simple verification of passed orthusfit object
}
