% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_s3.R
\name{verify.pibblefit}
\alias{verify.pibblefit}
\title{Simple verification of passed pibblefit object}
\usage{
\method{verify}{pibblefit}(m, ...)
}
\arguments{
\item{m}{an object of class pibblefit}

\item{...}{not used}
}
\value{
throws error if any verification tests fail
}
\description{
Simple verification of passed pibblefit object
}
