% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R
\name{verify}
\alias{verify}
\title{Generic method for verifying new objects}
\usage{
verify(m, ...)
}
\arguments{
\item{m}{object}

\item{...}{other arguments to be passed to verify}
}
\value{
throws error if verify test fails
}
\description{
Intended to be called internally by package or object creator
}
