% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheel1}
\alias{wheel1}
\title{wheel data set with inlier and outlier.}
\format{
A data frame with 1002 rows and 3 variables:
 \describe{
\item{x }{numerical variable}
\item{y }{numerical variable}
\item{type}{Type of a data point : Typical or Outlier}}
}
\usage{
wheel1
}
\description{
A bivariate dataset with an inlier and  anoutlier
}
\keyword{datasets}
