% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrelfit_methods.R
\name{access_dims}
\alias{access_dims}
\alias{ncategories}
\alias{nsamples}
\alias{ncovariates}
\title{Simple helper functions to access mongrel fit dimensions}
\usage{
ncategories(m)

nsamples(m)

ncovariates(m)
}
\arguments{
\item{m}{An object of class mongrelfit}
}
\value{
integer
}
\description{
Simple helper functions to access mongrel fit dimensions
}
\details{
An alternative approach to accessing these dimensions is to 
  access them directly from the mongrelfit object using list indexing. 
* \code{ncategories} is equivalent to \code{m$D}
* \code{nsamples} is equivalent to \code{m$N}
* \code{ncovariates} is equivalent to \code{m$Q}
}
\examples{
\dontrun{
m <- mongrel(Y, X)
ncategories(m)
nsamples(m)
ncovariates(m)
}
}
