% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R, R/strayfit_methods.R
\name{access_dims}
\alias{access_dims}
\alias{ncategories}
\alias{nsamples}
\alias{ncovariates}
\alias{niter}
\alias{ncategories.pibblefit}
\alias{nsamples.pibblefit}
\alias{ncovariates.pibblefit}
\alias{niter.pibblefit}
\alias{ncategories.orthusfit}
\alias{nsamples.orthusfit}
\alias{ncovariates.orthusfit}
\alias{niter.orthusfit}
\title{Generic method for accessing model fit dimensions}
\usage{
ncategories(m)

nsamples(m)

ncovariates(m)

niter(m)

\method{ncategories}{pibblefit}(m)

\method{nsamples}{pibblefit}(m)

\method{ncovariates}{pibblefit}(m)

\method{niter}{pibblefit}(m)

\method{ncategories}{orthusfit}(m)

\method{nsamples}{orthusfit}(m)

\method{ncovariates}{orthusfit}(m)

\method{niter}{orthusfit}(m)
}
\arguments{
\item{m}{An object of class pibblefit}
}
\value{
integer
}
\description{
Generic method for accessing model fit dimensions
}
\details{
An alternative approach to accessing these dimensions is to 
  access them directly from the pibblefit object using list indexing. 
* \code{ncategories} is equivalent to \code{m$D}
* \code{nsamples} is equivalent to \code{m$N}
* \code{ncovariates} is equivalent to \code{m$Q}
}
\examples{
\dontrun{
m <- pibble(Y, X)
ncategories(m)
nsamples(m)
ncovariates(m)
}
}
