% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_methods.R
\name{as.list.mongrelfit}
\alias{as.list.mongrelfit}
\title{Convert object of class mongrelfit to a list}
\usage{
\method{as.list}{mongrelfit}(m)
}
\arguments{
\item{m}{an object of class mongrelfit}
}
\description{
Convert object of class mongrelfit to a list
}
\examples{
\dontrun{
fit <- fit_mongrel(Y, X)
as.list(fit)
}
}
