% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{as.list.mongrelfit}
\alias{as.list.mongrelfit}
\title{Convert object of class mongrelfit to a list}
\usage{
\method{as.list}{mongrelfit}(x, ...)
}
\arguments{
\item{x}{an object of class mongrelfit}

\item{...}{currently unused}
}
\description{
Convert object of class mongrelfit to a list
}
\examples{
\dontrun{
fit <- pibble(Y, X)
as.list(fit)
}
}
