% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_HDoutliers.R
\name{display_HDoutliers}
\alias{display_HDoutliers}
\title{Display outliers with a scatterplot}
\usage{
display_HDoutliers(data, out, alpha = 1)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric
and/or categorical variables.}

\item{out}{A list containing output values produced by \code{\link[stray]{find_HDoutliers}}}

\item{alpha}{Alpha level of the points}
}
\description{
Animate a 2D tour path with a scatterplot to visualize outliers.
}
\examples{
data <- c(rnorm(100), 7, 7.5, rnorm(100, 20), 45)
out <- find_HDoutliers(data, method = "knn_maxdiff", knnsearchtype = "FNN_auto")
display_HDoutliers(data, out = out, alpha = out$out_scores)
\dontrun{
require(tourr)
outpoints <- matrix(rnorm(12, mean = 200), nrow = 2)
colnames(outpoints) <- colnames(flea[, -7])
data <- rbind(flea[, -7], outpoints)
outliers <- find_HDoutliers(data, method = "knn_maxdiff", knnsearchtype = "FNN_auto")
display_HDoutliers(data, outliers)
}
}
