% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{find_HDoutliers}
\alias{find_HDoutliers}
\title{Detect Anomalies in High Dimensional Data.}
\usage{
find_HDoutliers(data, maxrows = 10000, radius = NULL, alpha = 0.05,
  l = 0.2)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}

\item{maxrows}{If the number of observations is greater than \code{maxrows}, \code{outliers} reduces the
number used in k-nearest-neighbor computations to a set of \emph{exemplars}. The default value is 10000.}

\item{radius}{Threshold for determining membership in the \emph{exemplars}'s lists (used only when the
number of observations is greater than \eqn{maxrows}). An observation is added to an \emph{exemplars}'s lists
if its distance to that \emph{exemplar} is less than \code{radius}. The default value is \eqn{.1/(log n)^(1/p)},
 where \eqn{n} is the number of observations and \eqn{p} is the dimension of the data.}

\item{alpha}{Threshold for determining the cutoff for outliers. Observations are considered
outliers outliers if they fall in the \eqn{(1- alpha)} tail of the distribution of the nearest-neighbor
distances between exemplars.}

\item{l}{A user defined parameter to obtain a trade off between the false positives and the false negatives.}
}
\value{
The indexes of the observations determined to be outliers.
}
\description{
Detect anomalies in high dimensional data. This is a modification of
\code{\link[HDoutliers]{HDoutliers}}.
}
\details{
If the number of observations exceeds \code{maxrows}, the data is first partitioned into lists
associated with \emph{exemplars} and their \emph{members} within \code{radius} of each \emph{exemplar}, to
reduce the number of k-nearest neighbor computations required for outlier detection.
}
\examples{
require(ggplot2)
set.seed(1234)
data <- c(rnorm(1000, mean = -6), 0, rnorm(1000, mean = 6))
df <- data.frame( index = rep(0, length(data)), data = data)
data_plot <- ggplot(df, aes(x = data, y= index)) +
geom_point() +
xlab("x") +
ylab("") +
ggtitle("Original Data")
data_out <- find_HDoutliers(data)
output_plot  <- data_plot +
geom_point(data = df[data_out, ], aes(x=data, y = index),
colour = "red", size = 3) +
xlab("x") +
ylab("") +
ggtitle("Output")
gridExtra::grid.arrange(data_plot, output_plot )


set.seed(1234)
n <- 1000 # number of observations
nout <- 10 # number of outliers
typical_data <- tibble::as.tibble(matrix(rnorm(2*n), ncol = 2, byrow = TRUE))
out <- tibble::as.tibble(matrix(5*runif(2*nout,min=-5,max=5), ncol = 2, byrow = TRUE))
data <- dplyr::bind_rows(out, typical_data )
data_out <- find_HDoutliers(data)
data_plot <- ggplot(data, aes(x=V1, y= V2))+
geom_point() +
ggtitle("Original Data")+
theme(aspect.ratio = 1)
output_plot <- data_plot +
geom_point(data = data[data_out, ], aes(x=V1, y = V2),
colour = "red", size = 3) +
ggtitle("Output")
gridExtra::grid.arrange(data_plot, output_plot , nrow=1 )
}
\references{
{Wilkinson, L. (2016). Visualizing outliers.}
}
\seealso{
\code{\link[HDoutliers]{getHDmembers}},  \code{\link{get_outliers}}
}
