% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{get_leader_cluster}
\alias{get_leader_cluster}
\title{Form clusters of datapoints using Hartigan's Leader Algorithm.}
\usage{
get_leader_cluster(data, maxrows = 1000)
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}

\item{maxrows}{If the number of observations is greater than \code{maxrows}, \code{outliers} reduces the
number used in k-nearest-neighbor computations to a set of \emph{exemplars}. The default value is 1000.}
}
\value{
A list in which each component is a vector of
observation indexes. The first index in each list is the
index of the exemplar defining that list, and any remaining indexes are the associated members, within radius of the exemplar.
}
\description{
Form clusters of datapoints using Hartigan's Leader Algorithm. This is a modification of
\code{\link[HDoutliers]{getHDmembers}}.
}
\references{
{Hartigan, John A. "Clustering algorithms." (1975).}

{Kantardzic, Mehmed. Data mining: concepts, models, methods, and algorithms.
 John Wiley & Sons, 2011.}
}
\seealso{
\code{\link[HDoutliers]{getHDmembers}}
}
