% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels}
\alias{kernels}
\alias{RBF}
\alias{LINEAR}
\title{Multivariate RBF Kernel}
\usage{
RBF(X, sigma = 1, rho = median(dist))

LINEAR(X, sigma = 1, c = rowMeans(X))
}
\arguments{
\item{X}{covariate (dimension Q x N; i.e., covariates x samples)}

\item{sigma}{scalar parameter}

\item{rho}{scalar bandwidth parameter}

\item{c}{vector parameter defining intercept for linear kernel}
}
\value{
Gram Matrix (N x N) (e.g., the Kernel evaluated at 
each pair of points)
}
\description{
Designed to be partially specified. (see examples)
}
\details{
Gram matrix G is given by 

RBF:
\deqn{G = \sigma^2 * exp(-[(X-c)'(X-c)]/(s*\rho^2))}

LINEAR:
\deqn{G = \sigma^2*(X-c)'(X-c)}
}
\examples{
  # Create Partial for use with basset
  K <- function(X) RBF(X, 2, .2)
  
  # Example use
  X <- matrix(rnorm(10), 2, 5)
  G <- K(X)
  G # this is the gram matrix (the kernel evaluated on a finite set of points)
}
