% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmvgamma_deriv}
\alias{lmvgamma_deriv}
\title{Derivative of Log of Multivariate Gamma Function - Gamma_p(a)}
\usage{
lmvgamma_deriv(a, p)
}
\arguments{
\item{a}{defined by Gamma_p(a)}

\item{p}{defined by Gamma_p(a)}
}
\description{
Derivative of Log of Multivariate Gamma Function - Gamma_p(a)
}
\references{
https://en.wikipedia.org/wiki/Multivariate_gamma_function
}
