% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R, R/mongrelfit_methods.R
\name{ppc_summary}
\alias{ppc_summary}
\alias{ppc_summary.mongrelfit}
\title{Generic Method to Plot Posterior Predictive Summaries}
\usage{
ppc_summary(m, ...)

\method{ppc_summary}{mongrelfit}(m, ...)
}
\arguments{
\item{m}{model object}

\item{...}{other arguments to pass}
}
\value{
vector
}
\description{
Generic Method to Plot Posterior Predictive Summaries
}
