% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_s3.R, R/strayfit_methods.R
\name{ppc_summary}
\alias{ppc_summary}
\alias{ppc_summary.pibblefit}
\title{Generic Method to Plot Posterior Predictive Summaries}
\usage{
ppc_summary(m, ...)

\method{ppc_summary}{pibblefit}(m, from_scratch = FALSE, ...)
}
\arguments{
\item{m}{model object}

\item{...}{other arguments to pass}

\item{from_scratch}{should predictions of Y come from fitted Eta or from 
predictions of Eta from posterior of Lambda? (default: false)}
}
\value{
vector
}
\description{
Generic Method to Plot Posterior Predictive Summaries
}
