% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrel_methods.R
\name{print.mongrelfit}
\alias{print.mongrelfit}
\title{Print dimensions and coordinate system information for mongrelfit object.}
\usage{
\method{print}{mongrelfit}(m)
}
\arguments{
\item{m}{an object of class mongrelfit}
}
\description{
Print dimensions and coordinate system information for mongrelfit object.
}
\examples{
\dontrun{
fit <- fit_mongrel(Y, X)
print(fit)
}
}
\seealso{
\code{\link{summary.mongrelfit}} summarizes posterior intervals
}
