% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongrelfit_methods.R
\name{print.mongrelfit}
\alias{print.mongrelfit}
\title{Print dimensions and coordinate system information for mongrelfit object.}
\usage{
\method{print}{mongrelfit}(x, ...)
}
\arguments{
\item{x}{an object of class mongrelfit}

\item{...}{currently unused}
}
\description{
Print dimensions and coordinate system information for mongrelfit object.
}
\examples{
\dontrun{
fit <- mongrel(Y, X)
print(fit)
}
}
\seealso{
\code{\link{summary.mongrelfit}} summarizes posterior intervals
}
