% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayfit_methods.R
\name{sample_prior.mongrelfit}
\alias{sample_prior.mongrelfit}
\title{Sample from the prior distribution of mongrelfit object}
\usage{
\method{sample_prior}{mongrelfit}(m, n_samples = 2000, pars = c("Eta",
  "Lambda", "Sigma"), use_names = TRUE, ...)
}
\arguments{
\item{m}{object of class mongrelfit}

\item{n_samples}{number of samples to produce}

\item{pars}{parameters to sample}

\item{use_names}{should names be used if available}

\item{...}{currently ignored}
}
\description{
Note this can be used to sample from prior and then predict can
be called to get counts or LambdaX (\code{\link{predict.mongrelfit}})
}
\details{
Could be greatly speed up in the future if needed by sampling
directly from cholesky form of inverse wishart (currently implemented as 
header in this library - see MatDist.h).
}
\examples{
# Sample prior of already fitted  mongrelfit object
sim <- pibble_sim()
attach(sim)
fit <- pibble(Y, X)
sample_prior(fit)

# Sample prior as part of model fitting
m <- mongrelfit(N=N, D=D, Q=Q, iter=2000, upsilon=upsilon, 
                Xi=Xi, Gamma=Gamma, Theta=Theta, X=X, 
                coord_system="alr", alr_base=D)
m <- sample_prior(mongrelfit)
plot(m) # plot prior distribution (defaults to parameter Lambda) 
}
