% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_HDoutliers.R
\name{use_KNN}
\alias{use_KNN}
\title{Find outliers using kNN distance with maximum gap}
\usage{
use_KNN(data, alpha = 0.01, k = 10, knnsearchtype = c("kd_tree",
  "brute"))
}
\arguments{
\item{data}{A vector, matrix, or data frame consisting of numeric and/or categorical variables.}

\item{alpha}{Threshold for determining the cutoff for outliers. Observations are considered
outliers outliers if they fall in the \eqn{(1- alpha)} tail of the distribution of the nearest-neighbor
distances between exemplars.}

\item{k}{Number of neighbours considered.}

\item{knnsearchtype}{A character vector indicating the search type for k- nearest-neighbors.}
}
\value{
The indexes of the observations determined to be outliers and the outlying scores
}
\description{
Find outliers using kNN distance with maximum gap
}
